Fixed stack corruption bug in Double.toString(double). 	Change on 2013/10/31 by tball <tball@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=55933337

+++ jre_emul/android/libcore/luni/src/main/java/java/lang/RealToString.java
@@ -21,7 +21,7 @@ import libcore.math.MathUtils;
 
 /*-[
 extern void RealToString_bigIntDigitGenerator(
-    JavaLangRealToString *obj, long f, int e, BOOL isDenormalized, int p);
+    JavaLangRealToString *obj, long long f, int e, BOOL isDenormalized, int p);
 ]-*/
 
 final class RealToString {
